Rename directory to EventListener form Listener
