+++ DependencyInjection/Configuration.php
@@ -102,8 +102,8 @@ class Configuration implements ConfigurationInterface
                         ->scalarNode('json')->defaultValue('Symfony\Component\Serializer\Encoder\JsonEncoder')->end()
                         ->scalarNode('xml')->defaultValue('Symfony\Component\Serializer\Encoder\XmlEncoder')->end()
                         ->scalarNode('html')->defaultValue('FOS\RestBundle\Serializer\Encoder\HtmlEncoder')->end()
-                        ->scalarNode('body_listener')->defaultValue('FOS\RestBundle\Listener\RequestListener')->end()
-                        ->scalarNode('format_listener')->defaultValue('FOS\RestBundle\Listener\ControllerListener')->end()
+                        ->scalarNode('body_listener')->defaultValue('FOS\RestBundle\EventListener\RequestListener')->end()
+                        ->scalarNode('format_listener')->defaultValue('FOS\RestBundle\EventListener\ControllerListener')->end()
                     ->end()
                 ->end()
                 ->arrayNode('services')
