+++ Resources/config/frameworkextra_bundle.xml
@@ -6,7 +6,7 @@
 
     <parameters>
 
-        <parameter key="rest.template_annotation.class">FOS\RestBundle\Listener\AnnotationTemplateListener</parameter>
+        <parameter key="rest.template_annotation.class">FOS\RestBundle\EventListener\AnnotationTemplateListener</parameter>
 
     </parameters>
